[misc] Переместил задания для CSharp в отдельную папку
