+++ CSharpTasks/task-3/Calculator-1/Calculator 1/Calculator 1.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>Calculator_1</RootNamespace>
+  </PropertyGroup>
+
+</Project>
