+++ CSharpTasks/task-5/README.md
@@ -1,0 +1,38 @@
+# Задача: "CalculatorCMD"
+
+## Цель
+
+Разработать программу калькулятор. Консольное приложение, которое принимает как аргументы командной строки 2 числа и операцию, производит над ними действия, результат выводит на экран.
+
+## Поддерживаемые операции
+
+- сложение (+)
+- вычитание (-)
+- умножение (*)
+- деление (/)
+
+## Входные данные
+
+Два вещественных числа - операнды.
+Один символ - операция.
+
+Данные передаются как аргументы команданой строки: 
+
+```
+CalculatorCMD.exe 2 3 +
+```
+
+## Вывод
+
+Вывод производится на экран консоли: одно вещественное число - результат операции.
+
+## Пример
+
+Ввод:
+```
+CalculatorCMD.exe 15 2 /
+```
+Вывод:
+```
+7.5
+```
