+++ README.md
@@ -6,38 +6,38 @@
 
 ### Задача первая VolumeCalculator
 
-Условия задачи в папке [task-1](task-1/README.md)
+Условия задачи в папке [task-1](CSharpTasks/task-1/README.md)
 
 ## Задачи для Жени
 
 ### Задача первая Summator
 
-Условия задачи в папке [task-2](task-2/README.md)
+Условия задачи в папке [task-2](CSharpTasks/task-2/README.md)
 
 ### Задача вторая CalculatorCMD
 
-Условия задачи в папке [task-5](task-5/README.md)
+Условия задачи в папке [task-5](CSharpTasks/task-5/README.md)
 
 ### Задача третья PlantIndexer
 
-Условия задачи в папке [task-7](task-7/README.md)
+Условия задачи в папке [task-7](CSharpTasks/task-7/README.md)
 
 ### Упражнения на базовые вещи
 
-Условия этих упражнений в папке [Subtasks](Subtasks/README.md)
+Условия этих упражнений в папке [Subtasks](CSharpTasks/Subtasks/README.md)
 
 ## Задачи для Толи
 
 ### Задача первая Calculator
 
-Условия задачи в папке [task-3](task-3/README.md)
+Условия задачи в папке [task-3](CSharpTasks/task-3/README.md)
 
 ### Задача вторая ShapePacker
 
-Условия задачи в папке [task-4](task-4/README.md)
+Условия задачи в папке [task-4](CSharpTasks/task-4/README.md)
 
 ## Задачи для Димы
 
 ### Задача первая FileCalculator
 
-Условия задачи в папке [task-6](task-6/README.md)
+Условия задачи в папке [task-6](CSharpTasks/task-6/README.md)
