--- task-2/Summator/Summator.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Summator", "Summator\Summator.csproj", "{6746B19E-AE66-455C-9BA6-30AFDF3A8A3B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6746B19E-AE66-455C-9BA6-30AFDF3A8A3B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6746B19E-AE66-455C-9BA6-30AFDF3A8A3B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6746B19E-AE66-455C-9BA6-30AFDF3A8A3B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6746B19E-AE66-455C-9BA6-30AFDF3A8A3B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0A0F89D8-E26A-4C86-87B0-C802534AE99A}
-	EndGlobalSection
-EndGlobal
