--- task-3/Calculator-1/Calculator 1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator 1", "Calculator 1\Calculator 1.csproj", "{E2C0E346-73A5-4078-A6B5-F7C4B4E27B61}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E2C0E346-73A5-4078-A6B5-F7C4B4E27B61}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E2C0E346-73A5-4078-A6B5-F7C4B4E27B61}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E2C0E346-73A5-4078-A6B5-F7C4B4E27B61}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E2C0E346-73A5-4078-A6B5-F7C4B4E27B61}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CF117261-C41A-4B63-A24B-B2EDDD65B9A9}
-	EndGlobalSection
-EndGlobal
