--- task-4/README.md
@@ -1,65 +1,0 @@
-# Задача: "ShapePacker"
-
-## Цель
-
-Вычислить какие фигуры можно поместить в контейнер.
-
-## Описание 
-
-Программа является консольным приложением, которое получает на вход список с описанием трёхмерных геометрических фигур, а также объём контейнера. Программа вычисляет объём каждой фигуры и в том же порядке, в котором фигуры были переданы на вход, пытается разместить их в контейнере. 
-
-Если следующая фигура из списка уже не влезает в контейнер, программа должна продолжить проверять оставшиеся фигуры и пытаться добавить их в контейнер.
-
-Результатом работы программы является количество фигур, добавленных в контейнер и список их названий. Программа должна спросить у пользователя: выводить ли результат на экран или сохранить в файл. При выборе сохранить в файл: программа сохраняет результат в output.txt. После вывода, она завершает работу.
-
-## Входные данные
-
-Команда запускается с аргументом командным строки: путём до файла (абсолютным или относительным) с описаниями фигур.
-
-```
-ShapePacker.exe input.txt
-```
-
-Программа запрашивает объём контейнера у пользователя и читает ввод с консоли. Объём контейнера вводится как вещественное число.
-
-## Выходные данные
-
-Первая строка: число фигур, вошедших в контейнер.
-Следующие строки: названия фигур в том порядке, в котором они были добавлены в контейнер. По фигуре на строку.
-
-## Описание фигур
-
-Фигура | Формат описания | Формула расчёта объёма | Пример
------------- | -------------  | ------------- | ------------- 
-Cube | Тип Название H(Высота ребра) | ![Объём куба](https://i.imgur.com/9BOtDUI.png)| Cube Кубик-Рубик 3
-Cylinder| Тип Название H(высота цилиндра) R(радиус цилиндра) | ![Объём цилиндра](https://i.imgur.com/bqIIEVZ.png) | Cylinder Бочонок 3 2.4
-Pyramid | Тип Название H(высота пирамиды) A(сторона пирамиды A) B(сторона пирамиды B) | ![Объём пирамиды](https://i.imgur.com/fODIMNC.png) | Pyramid ГробницаМопса 7 2 3
-Sphere | Тип Название R(радиус сферы) | ![Объём сферы](https://i.imgur.com/ywBYcGA.png) | Sphere Дискошар 3
-
-Основной разделитель - пробел. В названиях фигур не допускаются пробелы! Все численные значения - вещественные.
-
-## Пример
-
-### Запуск программы
-
-```
-ShapePacker.exe someinputfile.txt
-```
-
-### Пример входного файла
-
-```
-Cube Шестигранник 2
-Sphere ВрекингБолл 100
-Cube КубикЛьда 3
-Pyramid ХанойскаяПирамидка 4 4 2
-Cylinder Шляпа 29 30
-```
-
-### Пример вывода
-
-```
-2
-Шестигранник
-КубикЛьда
-```
