--- task-5/CalculatorCMD/CalculatorCMD.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalculatorCMD", "CalculatorCMD\CalculatorCMD.csproj", "{280A0D1A-4289-479D-B259-CE2D0F45AEAB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{280A0D1A-4289-479D-B259-CE2D0F45AEAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{280A0D1A-4289-479D-B259-CE2D0F45AEAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{280A0D1A-4289-479D-B259-CE2D0F45AEAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{280A0D1A-4289-479D-B259-CE2D0F45AEAB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {21D4ADE6-4363-40A3-9E2B-D35307AE71EF}
-	EndGlobalSection
-EndGlobal
