--- task-5/CalculatorCMD/CalculatorCMD/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace CalculatorCMD
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-        }
-    }
-}
